Address selection & payment modal; phoneInspection -> int
Backend: change phoneInspection from sbyte/tinyint to int in DTO, model and DbContext mapping to ensure consistent typing between API and database.

Frontend: add address selection, saved-address lists and editable address forms, validation, and payment modal UI to Cart and ServiceRequest pages. Load addresses from API, allow choosing or adding new billing/delivery addresses, sync billingSameAsDelivery, and display total price. Add related CSS styles for modals, address displays and validation feedback. Also build and send updated repair payload (phoneInspection numeric flag, billing/delivery fields) when submitting a repair request.

+++ backend/PhoneScout_GitHub/DTOs/profileRepairPostDTO.cs
@@ -26,7 +26,7 @@ namespace PhoneScout_GitHub.DTOs
 
         public string manufacturerName { get; set; } = null!;
 
-        public sbyte phoneInspection { get; set; }
+        public int phoneInspection { get; set; }
     
         public string problemDescription { get; set; } = null!;
